# src/Happy-Spirit-Games_StoryTimeAPI_2.0.0_swagger.yaml
openapi: 3.0.1
info:
  title: StoryTime Service API
  description: >-
    Application programming interface (API) to the StoryTime service, supporting clients and consumers of interactive fiction stories where the player decides what path to take through the story.
  termsOfService: https://happyspiritgames.blog/terms-of-service/
  contact:
    name: Happy Spirit Games Support
    url: https://happyspiritgames.blog/terms-of-service/
    email: welcome@happyspiritgames.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "2.0.0"
servers:
  - url: https://virtserver.swaggerhub.com/Happy-Spirit-Games/StoryTimeReaderAPI/2.0.0
    description: API definition design and testing
  - url: https://hsg-storytime.herokuapp.com/api
    description: Production StoryTime service
  - url: https://hsg-storytime-staging.herokuapp.com/api
    description: Staging StoryTime service for pre-production testing
paths:
...
  /self/profile:
    get:
      tags:
        - player
      summary: Retrieves player profile information about the active user
      description: Retrieves player profile information about the active user, using
        authorization header to identify the player.
      operationId: getSelfProfile
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - player
      summary: Updates profile
      description: Updates player's own profile, using authorization header to identify the player. Only select properties are available for edit by the player.  Note that emailOptIn is a boolean, not a timestamp.  If true is passed while the stored field value is empty, the service will timestamp the field.  If true is passed while the field already has a timestamp, no change is made, preserving the existing timestamp.  If false is passed while the field has a value, the field will be nulled.  If false is passed while the field is null, no change is made.
      operationId: updateSelfProfile
      requestBody:
        description: Editable fields from PlayerProfile
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerProfileUpdate'
      responses:
        '202':
          description: Accepted changes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /codes/player-status:
    get:
      tags:
        - system
      summary: Returns all player status codes.
      description: Returns all player status codes for use in drop-downs and to look up display values.
      operationId: getPlayerStatusCodes
      responses:
        '200':
          description: Found codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeLookups'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/players/{playerId}:
    get:
      tags:
        - admin
      summary: Returns all players in the system
      description: Returns a list of players in the system with information that an administrator might want to see.
      operationId: getPlayer
      parameters:
        - name: playerId
          in: path
          description: player UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CodeLookup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        displayName:
          type: string
    CodeLookups:
      type: array
      items:
        $ref: '#/components/schemas/CodeLookup'
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Player:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nickname:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: integer
        emailOptInAt:
          type: string
          format: date-time
        authorOptInAt:
          type: string
          format: date-time
        penName:
          type: string
    PlayerProfileUpdate:
      type: object
      properties:
        nickname:
          type: string
        emailOptIn:
          type: boolean
...